Add DX12 bindings for nvngx-sys and nvngx crates
Extend api_gen to generate three separate FFI binding files:
- core_bindings.rs: API-agnostic types (enums, params, constants)
- vk_bindings.rs:   Vulkan-specific types and functions
- dx_bindings.rs:   DX12-specific types and functions

Hand-written opaque COM types (ID3D12Device, etc.) and pure-Rust
reimplementations of the static inline DX12 helpers live in nvngx-sys
behind #[cfg(windows)]. Vulkan bindings move from the crate root into
a `vk` module, mirroring the existing `dx` module.

In the high-level nvngx crate, API-agnostic types (FeatureHandle,
FeatureParameters, Feature, SuperSamplingOptimalSettings, FeatureRequirement,
CommonInfoStorage, ...) live in common.rs with function-pointer dispatch
for cleanup. The new dx module mirrors vk with System, SuperSamplingFeature,
and RayReconstructionFeature, accepting windows-rs COM types in its public
API.

Following the design integrated into master with `vk::get_feature_requirements`
+ `vk::System::get_capability_parameters(&self)`:

- `dx::get_feature_requirements` parallels its Vulkan sibling: an Init-free
  per-feature support check via `NVSDK_NGX_D3D12_GetFeatureRequirements`.
  Takes `&IDXGIAdapter` (windows-rs) per the upstream signature.
- `dx::System::get_capability_parameters(&self)` parallels its Vulkan
  sibling: the upstream `NVSDK_NGX_D3D12_GetCapabilityParameters` requires
  prior Init, and `&self` of `System` makes that requirement type-enforced.
- `FeatureRequirement` impl (is_supported / unsupported_reason /
  min_hw_architecture / min_os_version / check_supported) lives in common.rs
  so both backends share it. `min_os_version` returns `Result<&str>`,
  bubbling NUL-termination and UTF-8 decode errors.
- `CommonInfoStorage` lives in common.rs with `AsRef<NVSDK_NGX_FeatureCommonInfo>`,
  consumed by both `vk::System::new` / `vk::get_feature_requirements` and the
  DX12 equivalents via `common_info.map(CommonInfoStorage::new)`.

Drops the misleading `_vk` / `_dx` free functions on `FeatureParameters`
(`get_capability_parameters_*`, `supports_super_sampling_*`,
`supports_ray_reconstruction_*`): their names suggested "no Init required"
but they all called `GetCapabilityParameters` which is documented as
Init-required. Use `System::get_capability_parameters(&self)` once a
`System` exists, or the Init-free `get_feature_requirements` for pre-Init
support checks.

Mark `FeatureParameters::from_raw()` unsafe. Replace type aliases
(DlssFeature, RRFeature) with `#[doc(alias)]`. Rewrite
`FeatureParameters` Debug impl using `debug_struct` directly.

Lowercase `User32`/`Advapi32` → `user32`/`advapi32` in nvngx-sys/build.rs
so xwin-style cross-compile from Linux finds the import libraries on a
case-sensitive filesystem; MSVC link.exe is case-insensitive and unaffected.
