+++ crates/nvngx-sys/src/core_bindings.h
@@ -1,0 +1,10 @@
+#ifndef CORE_BINDINGS_H
+#define CORE_BINDINGS_H
+
+#include "../DLSS/include/nvsdk_ngx.h"
+#include "../DLSS/include/nvsdk_ngx_helpers.h"
+#include "../DLSS/include/nvsdk_ngx_helpers_dlssd.h"
+#include "../DLSS/include/nvsdk_ngx_defs_dlssd.h"
+#include "../DLSS/include/nvsdk_ngx_params_dlssd.h"
+
+#endif // CORE_BINDINGS_H
