+++ crates/nvngx-sys/src/lib.rs
@@ -5,17 +5,15 @@
 #![allow(non_snake_case)]
 #![allow(clippy::all)]
 
-use ash::vk::{
-    Buffer as VkBuffer, CommandBuffer as VkCommandBuffer, Device as VkDevice,
-    ExtensionProperties as VkExtensionProperties, Format as VkFormat, Image as VkImage,
-    ImageSubresourceRange as VkImageSubresourceRange, ImageView as VkImageView,
-    Instance as VkInstance, PFN_vkGetDeviceProcAddr, PFN_vkGetInstanceProcAddr,
-    PhysicalDevice as VkPhysicalDevice,
-};
 use libc::wchar_t;
 
-include!("bindings.rs");
+include!("core_bindings.rs");
 
 pub mod error;
 pub use error::*;
+#[cfg(windows)]
+pub mod dx;
+#[cfg(windows)]
+pub mod dx_helpers;
 pub mod helpers;
+pub mod vk;
