+++ crates/nvngx/Cargo.toml
@@ -18,3 +18,9 @@ log = "0.4.4"
 uuid = { version = "1", features = ["v4"] }
 derive_builder = "0.20"
 nvngx-sys = { version = "0.3.1",  path = "../nvngx-sys" }
+
+[target.'cfg(windows)'.dependencies]
+windows = { version = "0.61", features = [
+    "Win32_Graphics_Direct3D12",
+    "Win32_Graphics_Dxgi",
+] }
